requirements.txt
python-dotenv
google.generativeai
PyPDF2
streamlit 